CI - Put back the ci/ directory
Its missing from the `develop` branch currently, but it should be there, so putting it back.

Copy-pasted the `ci/` directory from the `main` branch.

My guess is this happened because of the `*.sh` rule in `.gitignore` so I added an exception in `.gitignore` to allow shell scripts in the `ci/` directory.