Backend: Refactor for Unit Tests & Reusable DB Calls (#147)
Refactoring code to make it more unit testable, which has the benefits of:
* The current dev process of deploying made it hard to debug because I
  don't have access to logs & its slower to iterate than locally.
  Writing code and running unit tests against it is much faster.
* If we can do some development locally, that means fewer deployments to "production"
* We can set up CI to run these unit tests and catch regressions

Refactored the DDB calls to be wrapped in a client, since I'm looking
at what we have to do for sorting and filtering and finding we'll need
to reuse similar DB calls. The abstraction also is generally a good idea
(in case for some reason we ever want to swap out the DB).